Created DataRetriever.java
