+++ src/main/java/seedu/navi/canteen/storage/DataRetriever.java
@@ -1,0 +1,223 @@
+package seedu.navi.canteen.storage;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class DataRetriever {
+
+    public static List<String> getCanteenData() {
+        List<String> data = new ArrayList<>();
+        data.add("Techno Edge | Western Food | F | F | F | F");
+        data.add("Techno Edge | Nasi Padang | F | T | F | F");
+        data.add("Techno Edge | Vegetarian | F | F | F | T");
+        data.add("Techno Edge | Mala Hotpot | T | F | F | F");
+        data.add("Techno Edge | Indian | F | F | F | F");
+        data.add("Techno Edge | Mixed Rice | F | F | F | F");
+        data.add("Techno Edge | Drinks & Snacks | F | F | F | F");
+        data.add("Techno Edge | Fruit Juice | F | F | F | F");
+        data.add("Techno Edge | Taiwan | F | F | F | F");
+        data.add("Techno Edge | Ramen & Soup | F | F  |F | F");
+        data.add("Techno Edge | Noodles | F | F | F | F");
+        data.add("Techno Edge | Chicken Rice | F | F | F | F");
+        data.add("Techno Edge | Bistro Box | T | F | T | F");
+        data.add("Frontier | Vegetarian | F | F | F | T");
+        data.add("Frontier | Korean | F | F | F | F");
+        data.add("Frontier | Indian | F | F | F | F");
+        data.add("Frontier | Noodles | F | F | F | F");
+        data.add("Frontier | Hotplates | F | F | F | F");
+        data.add("Frontier | Yong Tau Foo | F | F | F | F");
+        data.add("Frontier | Chef Wok | F | F | F | F");
+        data.add("Frontier | Muslim | F | T | F | F");
+        data.add("Frontier | Mala Hotpot | F | F | F | F");
+        data.add("Frontier | Western | F | T | F | F");
+        data.add("Frontier | Chicken Rice | F | F | F | F");
+        data.add("Frontier | Fruits & Juices | F | F | F | F");
+        data.add("Frontier | Gong Cha | F | F | F | F");
+        data.add("Frontier | Drinks & Snacks | F | F | F | F");
+        data.add("Frontier | Uncle Penyet Fusion | T | F | T | F");
+        data.add("Frontier | Li Ji Coffee House | F | F | T | F");
+        data.add("Frontier | Pasta Express | F | F | T | F");
+        data.add("Frontier | Taiwan Ichiban | F | F | T | F");
+        data.add("The Deck | Yong Tau Foo & Laksa | F | F | F | F");
+        data.add("The Deck | Chinese Cooked Food | F | F | F | F");
+        data.add("The Deck | Japanese | F | F | F | F");
+        data.add("The Deck | Snacks & Fried Kway Teow | F | F | F | F");
+        data.add("The Deck | Thai | F | F | F | F");
+        data.add("The Deck | Western | F | T | F | F");
+        data.add("The Deck | Noodle | F | F | F | F");
+        data.add("The Deck | Roasted Delights | F | F | F | F");
+        data.add("The Deck | Vegetarian | F | F | F | T");
+        data.add("The Deck | Indian Vegetarian | F | F | F | T");
+        data.add("The Deck | Muslim | F | T | F | F");
+        data.add("The Deck | Fresh Fruits & Juices | F | F | F | F");
+        data.add("The Deck | Bubble Tea | F | F | F | F");
+        data.add("The Deck | Drinks & Snacks | F | F | F | F");
+        data.add("The Deck | Liang Ban Gong Fu | F | F | T | F");
+        data.add("The Deck | Pasta Express | F | F | T | F");
+        data.add("The Deck | Mrs Hen | T | F | T | F");
+        data.add("Terrace | Local Noodles | F | F | F | F");
+        data.add("Terrace | Five Grains Noodles | F | F | F | F");
+        data.add("Terrace | Mixed Rice, Mala & Sichuan | F | F | F | F");
+        data.add("Terrace | Korean & Japanese | F | F | F | F");
+        data.add("Terrace | Drinks & Hot Foods | F | F | F | F");
+        data.add("PGP Canteen | PGP Indian Veg | F | F | F | T");
+        data.add("PGP Canteen | PGP Salad | F | F | F | F");
+        data.add("PGP Canteen | PGP Dough Magic | F | F | F | F");
+        data.add("PGP Canteen | PGP Fish Soup | F | F | F | F");
+        data.add("PGP Canteen | PGP Mala Hotpot | F | F | F | F");
+        data.add("PGP Canteen | PGP Vietnamese | F | F | F | F");
+        data.add("PGP Canteen | PGP Mini Wok | F | F | F | F");
+        data.add("PGP Canteen | PGP Western | F | F | F | F");
+        data.add("PGP Canteen | PGP Mixed Veg Rice | F | F | F | F");
+        data.add("PGP Canteen | PGP Chicken Rice | F | F | F | F");
+        data.add("PGP Canteen | PGP Drinks and Snacks | F | F | F | F");
+        data.add("Fine Food | Nasi Campur | T | F | F | F");
+        data.add("Fine Food | Japanese Fusion | T | F | T | F");
+        data.add("Fine Food | Duck & Chicken Rice | F | F | T | F");
+        data.add("Fine Food | Five Grains Bee Hoon | F | F | T | F");
+        data.add("Fine Food | Mala Hotpot | F | F | T | F");
+        data.add("Fine Food | Indian Cuisine | F | F | T | F");
+        data.add("Fine Food | Mixed Veg Rice | F | F | T | F");
+        data.add("Fine Food | Xiao Long Bao | F | F | T | F");
+        data.add("Fine Food | Hot Pot Snail Noodle | F | F | T | F");
+        data.add("Fine Food | Taiwan Cuisine | F | F | T | F");
+        data.add("Fine Food | Pasta/Fried Rice | F | F | T | F");
+        data.add("Fine Food | Western Cuisine | T | F | T | F");
+        data.add("Flavours | Dan Lao | F | F | F | F");
+        data.add("Flavours | Tenderfresh | T | F | F | F");
+        data.add("Flavours | Beverage | F | F | F | F");
+        data.add("Flavours | Yong Tau Foo | F | F | F | F");
+        data.add("Flavours | Chicken Rice | F | F | F | F");
+        data.add("Flavours | Mini Wok | F | F | F | F");
+        data.add("Flavours | Japanese | F | F | F | F");
+        data.add("Flavours | Taiwanese | F | F | F | F");
+        data.add("Flavours | Mala Hotpot | F | F | F | F");
+        data.add("Flavours | Noodles | F | F | F | F");
+        data.add("Flavours | Mixed Veg Rice | F | F | F | F");
+        data.add("Flavours | Gong Cha | F | F | F | F");
+        data.add("Flavours | Coconut Queen | F | F | F | F");
+        return data;
+    }
+
+    static List<String> getHostelData() {
+        List<String> data = new ArrayList<>();
+        data.add("Raffles: Techno Edge (550m) | Fine Food (650m) | " +
+                "Flavours (750m) | Frontier (900m) | The Deck (1200m) | " +
+                "Terrace (1800m) | PGP Canteen (2200m)");
+        data.add("Eusoff: The Deck (400m) | Terrace (550m) | " +
+                "Techno Edge (650m) | PGP Canteen (1500m) | " +
+                "Fine Food (1500m) | Flavours (1600m) | " +
+                "Frontier (1600m)");
+        data.add("Temasek: The Deck (400m) | Terrace (550m) | " +
+                "Techno Edge (650m) | PGP Canteen (1500m) | " +
+                "Fine Food (1500m) | Flavours (1600m) | " +
+                "Frontier (1600m)");
+        data.add("Sheares: Terrace (550m) | The Deck (600m) | " +
+                "PGP Canteen (1100m) | Techno Edge (1200m) | " +
+                "Frontier (1900m) | Fine Food (2100m) | " +
+                "Flavours (2200m)");
+        data.add("Kent Ridge: Terrace (400m) | The Deck (500m) | " +
+                "PGP Canteen (1100m) | Techno Edge (1100m) | " +
+                "Frontier (1800m) | Fine Food (2000m) | " +
+                "Flavours (2100m)");
+        data.add("King Edward: PGP Canteen (600m) | Terrace (1100m) | " +
+                "The Deck (1600m) | Frontier (850m) | " +
+                "Techno Edge (2000m) | Fine Food (2300m) | " +
+                "Flavours (2400m)");
+        data.add("PGP: PGP Canteen (270m) | Terrace (600m) | " +
+                "The Deck (1100m) | Frontier (1500m) | " +
+                "Techno Edge (1700m) | Fine Food (2600m) | " +
+                "Flavours (2700m)");
+        data.add("UTR: Fine Food (100m) | Flavours (200m) | " +
+                "Techno Edge (1300m) | Frontier (1600m) | " +
+                "The Deck (2000m) | Terrace (2600m) | " +
+                "PGP Canteen (3000m)");
+        data.add("RC4: Flavours (350m) | Fine Food (500m) | " +
+                "Techno Edge (1600m) | Frontier (1900m) | " +
+                "The Deck (2200m) | Terrace (2800m) | " +
+                "PGP Canteen (3200m)");
+        data.add("CAPT: Flavours (400m) | Fine Food (550m) | " +
+                "Techno Edge (1500m) | Frontier (1800m) | " +
+                "The Deck (2100m) | Terrace (2700m) | " +
+                "PGP Canteen (3100m)");
+        data.add("RVRC: Frontier (400m) | Techno Edge (850m) | " +
+                "Fine Food (1100m) | Flavours (1200m) | " +
+                "The Deck (1500m) | PGP Canteen (1800m) | " +
+                "Terrace (2100m)");
+        data.add("Tembusu: Flavours (200m) | Fine Food (250m) | " +
+                "Techno Edge (1400m) | Frontier (1700m) | " +
+                "The Deck (2100m) | Terrace (2700m) | " +
+                "PGP Canteen (3100m)");
+        data.add("Cinnamon: Flavours (200m) | Fine Food (250m) | " +
+                "Techno Edge (1400m) | Frontier (1700m) | " +
+                "The Deck (2100m) | Terrace (2700m) | " +
+                "PGP Canteen (3100m)");
+        return data;
+    }
+
+    static List<String> getOtherBuildingData() {
+        List<String> data = new ArrayList<>();
+        data.add("CELC: Techno Edge (400m) | The Deck (650m) | " +
+                "Terrace (850m) | Fine Food (1300m) | " +
+                "Flavours (1400m) | Frontier (1400m) | " +
+                "PGP Canteen (1800m)");
+        data.add("HSSML: Terrace (700m) | The Deck (500m) | " +
+                "PGP Canteen (950m) | Techno Edge (1100m) | " +
+                "Frontier (1750m) | Fine Food (2000m) | " +
+                "Flavours (2100m)");
+        data.add("CLB: Techno Edge (200m) | The Deck (750m) | " +
+                "Terrace (900m) | Fine Food (1000m) | " +
+                "Frontier (1100m) | Flavours (1100m) | " +
+                "PGP Canteen (1900m)");
+        data.add("MSL: Frontier (200m) | PGP Canteen (1200m) | " +
+                "Techno Edge (1300m) | Fine Food (1500m) | " +
+                "Flavours (1700m) | The Deck (2000m) | " +
+                "Terrace (2200m)");
+        data.add("UTown SRC: Fine Food (100m) | Flavours (200m) | " +
+                "Techno Edge (1400m) | Frontier (1700m) | " +
+                "The Deck (2100m) | Terrace (2300m) | " +
+                "PGP Canteen (3100m)");
+        data.add("UTown ERC: Flavours (250m) | Fine Food (300m) | " +
+                "Techno Edge (1300m) | Frontier (1600m) | " +
+                "The Deck (2000m) | Terrace (2200m) | " +
+                "PGP Canteen (3000m)");
+        data.add("University Hall: Frontier (350m) | Techno Edge (900m) | " +
+                "Fine Food (1100m) | Flavours (1200m) | " +
+                "The Deck (1600m) | PGP Canteen (1700m) | " +
+                "Terrace (1800m)");
+        data.add("MPSH: Fine Food (750m) | Techno Edge (850m) | " +
+                "Flavours (900m) | Frontier (1100m) | " +
+                "The Deck (1500m) | Terrace (2100m) | " +
+                "PGP Canteen (2500m)");
+        data.add("UHC: Frontier (550m) | Techno Edge (650m) | " +
+                "Fine Food (850m) | Flavours (1000m) | " +
+                "The Deck (1300m) | Terrace (1500m) | " +
+                "PGP Canteen (1900m)");
+        return data;
+    }
+
+    static List<String> getFacultyData() {
+        List<String> data = new ArrayList<>();
+        data.add("CDE: Techno Edge (400m) | Fine Food (850m) | " +
+                "Flavours (1000m) | The Deck (1100m) | " +
+                "Frontier (1300m) | Terrace (1700m) | " +
+                "PGP Canteen (2300m)");
+        data.add("FASS: The Deck (300m) | Terrace (450m) | " +
+                "Techno Edge (550m) | PGP Canteen (1400m) | " +
+                "Fine Food (1400m) | Frontier (1500m) | " +
+                "Flavours (1600m)");
+        data.add("SOC: Terrace (100m) | The Deck (220m) | " +
+                "PGP Canteen (1000m) | Techno Edge (1000m) | " +
+                "Fine Food (1900m) | Frontier (2000m) | " +
+                "Flavours (2000m)");
+        data.add("NBS: The Deck (400m) | Terrace (450m) | " +
+                "PGP Canteen (1000m) | Techno Edge (1000m) | " +
+                "Frontier (1800m) | Fine Food (1900m) | " +
+                "Flavours (2000m)");
+        data.add("Science: Frontier (220m) | Techno Edge (1100m) | " +
+                "Fine Food (1300m) | PGP Canteen (1400m) | " +
+                "Flavours (1500m) | Terrace (1700m) | " +
+                "The Deck (1800m)");
+        return data;
+    }
+}
